vqa_app/frontend/requirements.txt
Pillow
requests
streamlit 
Pillow